FossilOrigin-Name: 449403391791da033139e26ab25d2881125bef9493f45654d313b80e857a9ae1

+++ src/include/lib/Paheko/Entities/Web/Page.php
@@ -3,8 +3,10 @@
 namespace Paheko\Entities\Web;
 
 use Paheko\DB;
+use Paheko\DynamicList;
 use Paheko\Entity;
 use Paheko\Form;
+use Paheko\Plugins;
 use Paheko\Utils;
 use Paheko\Entities\Files\File;
 use Paheko\Files\Files;
@@ -12,6 +14,7 @@ use Paheko\Web\Render\Render;
 use Paheko\Web\Web;
 use Paheko\Web\Cache;
 use Paheko\UserTemplate\Modifiers;
+use Paheko\Users\DynamicFields;
 
 use KD2\DB\EntityManager as EM;
 
@@ -20,6 +23,7 @@ use const Paheko\{WWW_URL, ADMIN_URL};
 class Page extends Entity
 {
 	const NAME = 'Page du site web';
+	const PRIVATE_URL = '!web/?id=%d';
 
 	const TABLE = 'web_pages';
 
@@ -124,7 +128,7 @@ class Page extends Entity
 	{
 		$user_prefix = ADMIN_URL . 'web/?uri=';
 
-		$this->_html ??= Render::render($this->format, $this->dir(), $this->content, $user_prefix);
+		$this->_html ??= Render::render($this->format, $this->dir_path, $this->content, $user_prefix);
 
 		return $this->_html;
 	}
@@ -142,12 +146,54 @@ class Page extends Entity
 	public function preview(string $content): string
 	{
 		$user_prefix = ADMIN_URL . 'web/?uri=';
-		return Render::render($this->format, $this->dir(), $content, $user_prefix);
+		return Render::render($this->format, $this->dir_path, $content, $user_prefix);
 	}
 
 	public function path(): string
 	{
 		return $this->path;
+	}
+
+	public function listVersions(): DynamicList
+	{
+		$name_field = DynamicFields::getNameFieldsSQL('u');
+
+		$columns = [
+			'id' => ['select' => 'v.id'],
+			'id_user' => ['select' => 'v.id_user'],
+			'date' => [
+				'select' => 'v.date',
+				'label' => 'Date',
+			],
+			'author' => [
+				'label' => 'Auteur',
+				'select' => $name_field,
+			],
+			'size' => [
+				'label' => 'Longueur du texte',
+				'select' => 'v.size',
+			],
+			'changes' => [
+				'label' => 'Évolution',
+				'select' => 'v.changes',
+			],
+		];
+
+		$tables = 'web_pages_versions v
+			LEFT JOIN users u ON u.id = v.id_user';
+		$conditions = sprintf('v.id_page = %d', $this->id());
+		$list = new DynamicList($columns, $tables, $conditions);
+		$list->orderBy('id', true);
+
+		return $list;
+	}
+
+	public function getVersion(int $id): ?\stdClass
+	{
+		return DB::getInstance()->first('SELECT a.*,
+			IFNULL((SELECT content FROM web_pages_versions WHERE id_page = a.id_page AND id < a.id ORDER BY id DESC LIMIT 1), \'\') AS previous_content
+			FROM web_pages_versions a
+			WHERE a.id_page = ? AND a.id = ?;', $this->id(), $id) ?: null;
 	}
 
 	public function syncSearch(): void
@@ -160,6 +206,40 @@ class Page extends Entity
 		}
 
 		$this->dir()->indexForSearch(compact('content'), $this->title, 'text/html');
+	}
+
+	public function saveNewVersion(?int $user_id): bool
+	{
+		$content_modified = $this->isModified('content');
+		$prev_content = $this->getModifiedProperty('content');
+
+		$r = $this->save();
+
+		if ($content_modified) {
+			$db = DB::getInstance();
+			$l = mb_strlen($this->content);
+
+			$version = [
+				'id_user' => $user_id,
+				'id_page' => $this->id(),
+				'date'    => new \DateTime,
+				'content' => $this->content,
+				'size'    => $l,
+				'changes' => $l - mb_strlen($prev_content),
+			];
+
+			$db->insert('web_pages_versions', $version);
+			$version['id'] = $db->lastInsertId();
+
+			Plugins::fire('web.page.version.new', false, [
+				'entity'      => $this,
+				'content'     => $this->content,
+				'old_content' => $prev_content,
+				'version'     => (object) $version,
+			]);
+		}
+
+		return $r;
 	}
 
 	public function save(bool $selfcheck = true): bool
@@ -175,12 +255,11 @@ class Page extends Entity
 			$this->set('modified', new \DateTime);
 		}
 
-		$content_modified = $this->isModified('content') || $this->isModified('format');
+		$update_search = $this->isModified('content') || $this->isModified('format');
 
-		Files::ensureDirectoryExists($this->dir_path);
 		parent::save($selfcheck);
 
-		if ($content_modified) {
+		if ($update_search) {
 			$this->syncSearch();
 		}
 
@@ -206,12 +285,14 @@ class Page extends Entity
 	{
 		$dir = $this->dir();
 
-		if ($dir) {
+		$r = parent::delete();
+
+		if ($r && $dir) {
 			$dir->delete();
 		}
 
 		Cache::clear();
-		return parent::delete();
+		return $r;
 	}
 
 	public function selfCheck(): void
@@ -226,7 +307,6 @@ class Page extends Entity
 		$this->assert(trim($this->uri) !== '', 'L\'URI ne peut rester vide');
 		$this->assert(strlen($this->uri) <= 150, 'L\'URI ne peut faire plus de 150 caractères');
 		$this->assert($this->path !== $this->parent, 'Invalid parent page');
-		$this->assert($this->parent === null || $db->test(self::TABLE, 'path = ?', $this->parent), 'Page parent inexistante');
 
 		$this->assert(!$this->exists() || !$db->test(self::TABLE, 'uri = ? AND id != ?', $this->uri, $this->id()), 'Cette adresse URI est déjà utilisée par une autre page, merci d\'en choisir une autre : ' . $this->uri, self::DUPLICATE_URI_ERROR);
 		$this->assert($this->exists() || !$db->test(self::TABLE, 'uri = ?', $this->uri), 'Cette adresse URI est déjà utilisée par une autre page, merci d\'en choisir une autre : ' . $this->uri, self::DUPLICATE_URI_ERROR);
@@ -290,14 +370,14 @@ class Page extends Entity
 	public function getBreadcrumbs(): array
 	{
 		$sql = '
-			WITH RECURSIVE parents(title, parent, path, level) AS (
-				SELECT title, parent, path, 1 FROM web_pages WHERE id = ?
+			WITH RECURSIVE parents(title, parent, path, id, level) AS (
+				SELECT title, parent, path, id, 1 FROM web_pages WHERE id = ?
 				UNION ALL
-				SELECT p.title, p.parent, p.path, level + 1
+				SELECT p.title, p.parent, p.path, p.id, level + 1
 				FROM web_pages p
 					JOIN parents ON parents.parent = p.path
 			)
-			SELECT path, title FROM parents ORDER BY level DESC;';
+			SELECT id, title FROM parents ORDER BY level DESC;';
 		return DB::getInstance()->getAssoc($sql, $this->id());
 	}
 
@@ -322,7 +402,7 @@ class Page extends Entity
 	{
 		if (null === $this->_tagged_attachments) {
 			$this->render();
-			$this->_tagged_attachments = Render::listAttachments($this->dir());
+			$this->_tagged_attachments = Render::listAttachments($this->dir_path);
 		}
 
 		return $this->_tagged_attachments;
@@ -395,7 +475,7 @@ class Page extends Entity
 			return [];
 		}
 
-		$html = Render::render($this->format, $this->dir(), $this->content);
+		$html = Render::render($this->format, $this->dir_path, $this->content);
 		preg_match_all('/<a[^>]+href=["\']([^"\']+)["\']/', $html, $match, PREG_PATTERN_ORDER);
 		$errors = [];
 
@@ -422,25 +502,17 @@ class Page extends Entity
 		return array_unique($errors);
 	}
 
-	public function checkRealType(): int
+	public function hasSubPages(): bool
 	{
-		// Make sure this is actually not a category
-		foreach (Files::list($this->dir_path) as $subfile) {
-			if ($subfile->type == File::TYPE_DIRECTORY) {
-				return self::TYPE_CATEGORY;
-			}
-		}
-
-		// No subdirectory found: this is a single page
-		return self::TYPE_PAGE;
+		return DB::getInstance()->test('web_pages', 'parent = ?', $this->path);
 	}
 
 	public function toggleType(): void
 	{
-		$real_type = $this->checkRealType();
+		$has_sub_pages = $this->hasSubPages();
 
-		if ($real_type == self::TYPE_CATEGORY) {
-			$this->set('type', $real_type);
+		if ($has_sub_pages) {
+			$this->set('type', self::TYPE_CATEGORY);
 		}
 		elseif ($this->type == self::TYPE_CATEGORY) {
 			$this->set('type', self::TYPE_PAGE);
